+++ vscode-dotnet-runtime-library/src/Utils/CommandExecutor.ts
@@ -17,7 +17,9 @@ import {
     CommandExecutionStdError,
     CommandExecutionStdOut,
     CommandExecutionUnderSudoEvent,
+    CommandExecutionUserAskDialogueEvent,
     CommandExecutionUserCompletedDialogueEvent,
+    CommandExecutionUserRejectedPasswordRequest,
     DotnetAlternativeCommandFoundEvent,
     DotnetCommandNotFoundEvent,
     DotnetWSLSecurityError
@@ -89,6 +91,8 @@ Please install the .NET SDK manually by following https://learn.microsoft.com/en
             let sanitizedCallerName = this.context?.acquisitionContext?.requestingExtensionId?.replace(/[^0-9a-z]/gi, ''); // Remove non-alphanumerics per OS requirements
             sanitizedCallerName = sanitizedCallerName?.substring(0, 69); // 70 Characters is the maximum limit we can use for the prompt.
             const options = { name: `${sanitizedCallerName ?? '.NET Install Tool'}` };
+
+            this.context?.eventStream.post(new CommandExecutionUserAskDialogueEvent(`Prompting user for command ${fullCommandString} under sudo.`));
             exec((fullCommandString), options, (error?: any, stdout?: any, stderr?: any) =>
             {
                 let commandResultString = '';
@@ -111,6 +115,14 @@ ${stderr}`));
                     this.context?.eventStream.post(new CommandExecutionUserCompletedDialogueEvent(`The command ${fullCommandString} failed to run under sudo.`));
                     if(terminalFailure)
                     {
+                        if(error.code === 126)
+                        {
+                            const err = new CommandExecutionUserRejectedPasswordRequest(new Error(`Cancelling .NET Install, as command ${fullCommandString} failed.
+The user refused the password prompt.`),
+                                getInstallKeyFromContext(this.context?.acquisitionContext!));
+                            this.context?.eventStream.post(err);
+                            reject(err.error);
+                        }
                         reject(error);
                     }
                     else