+++ vscode-dotnet-runtime-library/src/test/unit/RedHatDistroTests.test.ts
@@ -53,7 +53,7 @@ suite('Red Hat For Linux Distro Logic Unit Tests', () =>
         if(shouldRun)
         {
             const distroFeedDir = await provider.getExpectedDotnetDistroFeedInstallationDirectory();
-            assert.equal(distroFeedDir, '/usr/lib64/dotnet/dotnet');
+            assert.equal(distroFeedDir, '/usr/lib64/dotnet');
         }
     }).timeout(standardTimeoutTime);